address final comments
