+++ src/services/behaviors/index.js
@@ -8,10 +8,16 @@
  *
  */
 
-import type { BehaviorRegistry, HvBehavior } from 'hyperview/src/types';
+import type {
+  BehaviorRegistry,
+  Document,
+  Element,
+  HvBehavior,
+} from 'hyperview/src/types';
 import HvAlert from 'hyperview/src/behaviors/hv-alert';
 import HvShare from 'hyperview/src/behaviors/hv-share';
 import HvToggle from 'hyperview/src/behaviors/hv-toggle';
+import { shallowCloneToRoot } from 'hyperview/src/services';
 
 const HYPERVIEW_BEHAVIORS = [HvAlert, HvShare, HvToggle];
 
@@ -23,3 +29,52 @@ export const getRegistry = (behaviors: HvBehavior[] = []): BehaviorRegistry =>
     }),
     {},
   );
+
+/**
+ * The elements in the Document with the given ids will be set to
+ * visible or hidden based on showIndicators.
+ * Returns a new Document object with the toggled indicators.
+ */
+export const toggleIndicators = (
+  ids: Array<string>,
+  showIndicators: boolean,
+  root: Document,
+): Document =>
+  ids.reduce((newRoot, id) => {
+    const indicatorElement: ?Element = root.getElementById(id);
+    if (!indicatorElement) {
+      return newRoot;
+    }
+    indicatorElement.setAttribute('hide', showIndicators ? 'false' : 'true');
+    return shallowCloneToRoot(indicatorElement);
+  }, root);
+
+/**
+ * Returns a new Document object that shows the "show" indicators
+ * and hides the "hide" indicators.
+ */
+export const setIndicatorsBeforeLoad = (
+  showIndicatorIds: Array<string>,
+  hideIndicatorIds: Array<string>,
+  root: Document,
+): Document => {
+  let newRoot: Document = root;
+  newRoot = toggleIndicators(showIndicatorIds, true, newRoot);
+  newRoot = toggleIndicators(hideIndicatorIds, false, newRoot);
+  return newRoot;
+};
+
+/**
+ * Returns a new Document object that hides the "show" indicators
+ * and shows the "hide" indicators.
+ */
+export const setIndicatorsAfterLoad = (
+  showIndicatorIds: Array<string>,
+  hideIndicatorIds: Array<string>,
+  root: Document,
+): Document => {
+  let newRoot: Document = root;
+  newRoot = toggleIndicators(showIndicatorIds, false, newRoot);
+  newRoot = toggleIndicators(hideIndicatorIds, true, newRoot);
+  return newRoot;
+};
